+++ src/main/java/top/code666/huobi/common/entity/marketdata/Trade1.java
@@ -14,9 +14,9 @@ import java.io.Serializable;
 @Data
 public class Trade1 implements Serializable {
     private static final long serialVersionUID = -3763062828889276757L;
-    private int id; // 唯一交易id
+    private long id; // 唯一交易id
     private float amount; // 以基础币种为单位的交易量
     private float price; // 以报价币种为单位的成交价格
-    private int ts; // 北京时间的时间戳，单位为毫秒
+    private long ts; // 北京时间的时间戳，单位为毫秒
     private String direction; // 交易方向:"buy" "sell"
 }
